RAG性能调优/34-CRAG纠正性检索增强/src/requirements.txt
llama-index
llama-index-llms-openai-like
llama-index-embeddings-dashscope
llama-index-tools-tavily-research
nest-asyncio
python-dotenv